Make selections work
Added functions to grab options from chart, another to take
those options and return a selection, and another to set the
text in the selection element.

Had to remove the onblur method because it was unloading
right after clicking the "make selection" button, and then
there weren't any options to select from.